% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatic_outliers.R
\name{climatic_outliers}
\alias{climatic_outliers}
\title{Climatic outliers test}
\usage{
climatic_outliers(xts_vector, ext_lim_factor = 4)
}
\arguments{
\item{xts_vector}{An xts object with one column 
(if multicolumn, only the first column will be considered).}

\item{ext_lim_factor}{Number of IQRs that must be exceeded.}
}
\value{
A list of two xts objects (data with outliers removed and flags).
}
\description{
Flag statistical outliers defined from the interquartile range (IQR).
}
\examples{
out_list <- climatic_outliers(Bern$data$Log_1)

}
