% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gross_errors.R
\name{gross_errors}
\alias{gross_errors}
\title{Gross errors test}
\usage{
gross_errors(xts_vector)
}
\arguments{
\item{xts_vector}{An xts object with one column 
(if multicolumn, only the first column will be considered).}
}
\value{
A list of two xts objects (data with outliers removed and flags).
}
\description{
Flag impossible values.
}
\examples{
out_list <- gross_errors(Bern$data$Log_1)

}
