% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_of_range.R
\name{out_of_range}
\alias{out_of_range}
\title{Out of range test}
\usage{
out_of_range(
  xts_vector,
  season_thresholds = list(winter = list(min_val = -24.44, max_val = 23.54), spring =
    list(min_val = -21.05, max_val = 38.08), summer = list(min_val = 7.85, max_val =
    46.26), autumn = list(min_val = -8.42, max_val = 40.87))
)
}
\arguments{
\item{xts_vector}{An xts object with one column 
(if multicolumn, only the first column will be considered).}

\item{season_thresholds}{A list of lists of thresholds (one list per season, default values are for Bern).}
}
\value{
A list of two xts objects (data with outliers removed and flags).
}
\description{
Flag the values that exceed custom thresholds based on climatology.
}
\examples{
out_list <- out_of_range(Bern$data$Log_1)

}
