% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_consistency.R
\name{spatial_consistency}
\alias{spatial_consistency}
\title{Spatial consistency test}
\usage{
spatial_consistency(
  timeseries_df,
  metadata,
  threshold_multiplier = 3,
  weight_decay = NULL,
  id_col = "ID",
  distance_threshold_method = "std",
  min_neighbors = 2,
  min_threshold = 0.5
)
}
\arguments{
\item{timeseries_df}{An xts object with the entire data set or a list of xts objects to be combined.}

\item{metadata}{An xts object with spatial metadata including land use.}

\item{threshold_multiplier}{Defines by how much to inflate differences (the higher, the more flags).}

\item{weight_decay}{Defines decrease of weights with distance. Use median distance if set to NULL.}

\item{id_col}{Name of column with station IDs}

\item{distance_threshold_method}{"std", "mean", or a numeric value.}

\item{min_neighbors}{Minimum number of neighbors}

\item{min_threshold}{Defines threshold to flag data (the lower, the more flags).}
}
\value{
A list of two xts objects (data with outliers removed and flags).
}
\description{
Flag values based on neighboring stations with same land use category.
}
\examples{
out_list <- spatial_consistency(Bern$data[1,], Bern$xyz)

}
