% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatiotemporal_consistency.R
\name{spatiotemporal_consistency}
\alias{spatiotemporal_consistency}
\title{Spatio-temporal consistency test}
\usage{
spatiotemporal_consistency(
  xy_target,
  xts_database,
  xyz_metadata,
  params = list(lmt_xy_ = 2500, lmt_n_ = 5, n_nearby_ = 5),
  threshold = 0.99
)
}
\arguments{
\item{xy_target}{ID of target station.}

\item{xts_database}{An xts object with temperature data.}

\item{xyz_metadata}{An xts object with spatial metadata.}

\item{params}{List of parameters lmt_xy_, lmt_n_, n_nearby_.}

\item{threshold}{Quantile threshold}
}
\value{
A list of two xts objects (data with outliers removed and flags for target station)
and a vector of flagged times.
}
\description{
Flag values based on neighboring stations and time evolution simultaneously.
}
\examples{
out_list <- spatiotemporal_consistency("Log_1", Bern$data, Bern$xyz)

}
