% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_persistence.R
\name{temporal_persistence}
\alias{temporal_persistence}
\title{Temporal persistence test}
\usage{
temporal_persistence(
  xts_vector,
  window_size = 180,
  by = "10 min",
  na_tolerance = 5,
  min_non_na = 1,
  threshold = 0.001
)
}
\arguments{
\item{xts_vector}{An xts object with one column 
(if multicolumn, only the first column will be considered).}

\item{window_size}{Time window length in minutes.}

\item{by}{Time resolution (string such as "10 min").}

\item{na_tolerance}{Maximum number of NAs allowed in the window.}

\item{min_non_na}{Minimum number on non-NA values in the window.}

\item{threshold}{Minimum standard deviation allowed in the window.}
}
\value{
A list of two xts objects (data with outliers removed and flags).
}
\description{
Flag equal or nearly equal consecutive values within a time window.
}
\examples{
out_list <- temporal_persistence(Bern$data$Log_1)

}
