% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_consistency.R
\name{time_consistency}
\alias{time_consistency}
\title{Time consistency test}
\usage{
time_consistency(xts_vector, dt = 3, diff = 6)
}
\arguments{
\item{xts_vector}{An xts object with one column 
(if multicolumn, only the first column will be considered).}

\item{dt}{Time window half-width (number of time steps).}

\item{diff}{Maximum difference from running median allowed.}
}
\value{
A list of two xts objects (data with outliers removed and flags).
}
\description{
Flag too large deviations from a running median.
}
\examples{
out_list <- time_consistency(Bern$data$Log_1)

}
